Fixed bug where ready for publish is not set when running remote.
