+++ src/Altinn.Correspondence.Application/MalwareScanResult/MalwareScanResultHandler.cs
@@ -5,6 +5,7 @@ using Altinn.Correspondence.Core.Models.Enums;
 using Altinn.Correspondence.Core.Repositories;
 using Altinn.Correspondence.Core.Services;
 using Altinn.Correspondence.Core.Services.Enums;
+using Altinn.Correspondence.Persistence.Repositories;
 using Microsoft.Extensions.Logging;
 using OneOf;
 
@@ -15,6 +16,8 @@ public class MalwareScanResultHandler(
     IAttachmentStatusRepository attachmentStatusRepository,
     IEventBus eventBus,
     UploadHelper uploadHelper,
+    ICorrespondenceRepository correspondenceRepository,
+    ICorrespondenceStatusRepository correspondenceStatusRepository,
     ILogger<MalwareScanResultHandler> logger) : IHandler<ScanResultData, Task>
 {
     private readonly IAttachmentRepository _attachmentRepository = attachmentRepository;
@@ -67,5 +70,36 @@ public class MalwareScanResultHandler(
             _logger.LogWarning("Malicious result for {fileTransferId} with result type {scanResultType}", attachmentId, data.ScanResultType);
             return Task.CompletedTask;
         }
+    }
+
+    public async Task CheckCorrespondenceStatusesAfterDeleteAndPublish(Guid attachmentId, CancellationToken cancellationToken)
+    {
+        var attachment = await _attachmentRepository.GetAttachmentById(attachmentId, true, cancellationToken);
+        if (attachment == null)
+        {
+            return;
+        }
+
+        var correspondences = await correspondenceRepository.GetNonPublishedCorrespondencesByAttachmentId(attachment.Id, cancellationToken);
+        if (correspondences.Count == 0)
+        {
+            return;
+        }
+
+        var list = new List<CorrespondenceStatusEntity>();
+        foreach (var correspondenceId in correspondences)
+        {
+            list.Add(
+                new CorrespondenceStatusEntity
+                {
+                    CorrespondenceId = correspondenceId,
+                    Status = CorrespondenceStatus.Published,
+                    StatusChanged = DateTime.UtcNow,
+                    StatusText = CorrespondenceStatus.Published.ToString()
+                }
+            );
+        }
+        await correspondenceStatusRepository.AddCorrespondenceStatuses(list, cancellationToken);
+        return;
     }
 }
